Feat: Home page

+++ CustomKicks/app/Http/Controllers/AdminCustomizationController.php
@@ -20,6 +20,9 @@ class AdminCustomizationController extends Controller
 
     public function edit($id)
     {
+        if (auth()->user()->role !== 'admin') {
+            abort(403, 'No tienes permisos para acceder a esta página.');
+        }
         $customization = Customization::findOrFail($id);
 
         return view('admin.editCustomization')->with('customization', $customization);
@@ -27,12 +30,14 @@ class AdminCustomizationController extends Controller
 
     public function update(Request $request, $id)
     {
+        if (auth()->user()->role !== 'admin') {
+            abort(403, 'No tienes permisos para acceder a esta página.');
+        }
         $customization = Customization::findOrFail($id);
 
-        $customization->setColor($request->input('color'));
-        $customization->setDesign($request->input('design'));
-        $customization->setPattern($request->input('pattern'));
-        $customization->setImage($request->input('image'));
+        $customization->color = $request->input('color');
+        $customization->design = $request->input('design');
+        $customization->pattern = $request->input('pattern');
         $customization->save();
 
         return redirect()->route('admin.customizations.dashboard')->with('success', 'Customización actualizada correctamente.');
@@ -40,6 +45,9 @@ class AdminCustomizationController extends Controller
 
     public function delete($id)
     {
+        if (auth()->user()->role !== 'admin') {
+            abort(403, 'No tienes permisos para acceder a esta página.');
+        }
         Customization::destroy($id);
 
         return redirect()->route('admin.customizations.dashboard')->with('success', 'Customización eliminada correctamente.');
@@ -47,11 +55,17 @@ class AdminCustomizationController extends Controller
 
     public function add()
     {
+        if (auth()->user()->role !== 'admin') {
+            abort(403, 'No tienes permisos para acceder a esta página.');
+        }
         return view('admin.addCustomization');
     }
 
     public function store(Request $request)
     {
+        if (auth()->user()->role !== 'admin') {
+            abort(403, 'No tienes permisos para acceder a esta página.');
+        }
 
         Customization::validations($request);
 
@@ -59,13 +73,6 @@ class AdminCustomizationController extends Controller
         $customization->setColor($request->input('color'));
         $customization->setDesign($request->input('design'));
         $customization->setPattern($request->input('pattern'));
-        $customization->setImage($request->input('image'));
-
-        if ($request->hasFile('image')) {
-            $path = $request->file('image')->store('customizations', 'public');
-            $customization->setImage($path);
-        }
-
         $customization->save();
 
         return redirect()->route('admin.customizations.dashboard')->with('success', 'Customización agregada correctamente.');
